+++ packages/solidart_hooks/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- **FEAT**: Added `useResource`, `useResourceStream`, `useListSignal`, `useSetSignal` and `useMapSignal` hooks.
+- **CHORE**: Export `flutter_solidart` package.
+
 ## 1.0.0+1
 
 - **CHORE**: Move example to `packages/solidart_hooks/example` to be compatible with pub.dev requirements.